% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_counter.R
\name{datagen.counter}
\alias{datagen.counter}
\title{Data generation function for counter example}
\usage{
datagen.counter(n)
}
\arguments{
\item{n}{sample size}
}
\value{
The resulting data has 4 lists: X, Y, M, V1
}
\description{
This function generates the simulation data (exposure X, outcome Y, potential mediators M, and covariate V1) for counter example.

The number of potential mediators is set to be p = 19.

For each subject i = 1, ..., n:
\itemize{
\item \eqn{X_i \stackrel{i.i.d}{\sim} N(0, 1)}
\item \eqn{{V_1}_i \stackrel{i.i.d}{\sim} N(0, 1)}
\item \eqn{cov(X, V_1) = 0.8}
\item \eqn{M_{i,k} = \alpha_k X_i + {V_1}_i {\eta_1}_k + e_{1_{i,k}}}, where \eqn{e_{1_{i,k}} \stackrel{i.i.d}{\sim} N(0, 1), k = 1, ..., p}, and \eqn{cov(e_{1_{i,k}},e_{1_{i,j}} = 0)} for k not equal j
\item \eqn{Y_i = X_i + \sum_k \beta_{k} M_{i,k} + e_{2_{i}}}, where \eqn{e_{2_{i}} \stackrel{i.i.d}{\sim} N(0, 1)}
}

We let \eqn{\alpha_1, \cdots, \alpha_{10}} be randomly from the uniform distribution from -1 to 1 and \eqn{\alpha_{11}, \cdots, \alpha_{19}} be 0.

We let \eqn{{\eta_1}_{11} = 1} and the rest of the \eqn{\eta_1} values be 0.

We let \eqn{\beta_1, \cdots, \beta{9} = 0} and \eqn{\beta_{10}, \cdots, \beta{19}} be either 1 or -1, randomly.

M10 is the true mediators (i.e., having non-zero alpha and beta coefficients).
}
\examples{
set.seed(23)
data = datagen.counter(n = 1000)
head(data$X)
head(data$Y)
head(data$M)
head(data$V1)
}
