% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen}
\alias{datagen}
\title{Data generation function}
\usage{
datagen(n, K, a, b)
}
\arguments{
\item{n}{sample size}

\item{K}{number of potential mediators}

\item{a}{effect size for X-M path}

\item{b}{effect size for M-Y path}
}
\value{
The resulting dataset has 3 lists: X, Y and M
}
\description{
This function generates the simulation data (exposure X, outcome Y, and potential mediators M).
}
\examples{
datagen(100, 50, 0.5, 0.3)
}
