% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediationFDR.R
\name{MediationFDR}
\alias{MediationFDR}
\title{Mediator selection function}
\usage{
MediationFDR(
  X,
  Y,
  M,
  V1 = NULL,
  V2 = NULL,
  q1 = 0.025,
  q2 = 0.025,
  f_func = "Diff",
  correction_method = "holm",
  weighted = FALSE,
  binaryOutcome = FALSE
)
}
\arguments{
\item{X}{exposure of interest}

\item{Y}{outcome of interest}

\item{M}{potential mediators}

\item{V1}{confounding variables in X-M path}

\item{V2}{confounding variables in M-Y path}

\item{q1}{FWER level in step 1}

\item{q2}{FDR controlled level in step 4}

\item{f_func}{function form of W in step 4}

\item{correction_method}{multiple comparison correction method in step 1}

\item{weighted}{whether we weight the Z statistics in step 3}

\item{binaryOutcome}{whether the outcome is binary}
}
\value{
selected mediators
}
\description{
This function selects the mediators while controlling for FDR.
}
\examples{
set.seed(20205)
data = datagen(n = 1000, p = 100, a = 0.5, b = 0.3, rho = 0.1)
MediationFDR(X = data$X, Y = data$Y, M = data$M, V1 = NULL, V2 = NULL, q1 = 0.025, q2 = 0.025, f_func = "Diff", correction_method = "holm", weighted = FALSE)
}
