% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen}
\alias{datagen}
\title{Data generation function}
\usage{
datagen(n, K, a, b)
}
\arguments{
\item{n}{sample size}

\item{K}{number of potential mediators}

\item{a}{effect size for X-M path (i.e., non-zero alpha coefficient)}

\item{b}{effect size for M-Y path (i.e., non-zero beta coefficient)}
}
\value{
The resulting dataset has 3 lists: X, Y and M
}
\description{
This function generates the simulation data (exposure X, outcome Y, and potential mediators M).

For each subject i = 1, ..., n:
\itemize{
\item \eqn{X_i \stackrel{i.i.d}{\sim} Bernoulli(p = 0.5)}
\item \eqn{M_{i,k} = \alpha_k X_i + e_{1_{i,k}}}, where \eqn{e_{1_{i,k}} \stackrel{i.i.d}{\sim} N(0, 1), k = 1, ..., K}
\item \eqn{Y_i = X_i + \sum_k b_{k} M_{i,k} + e_{2_{i}}}, where \eqn{e_{2_{i}} \stackrel{i.i.d}{\sim} N(0, 1)}
}

The 6th-10th M variables (M6,M7,M8,M9,M10) are set to be the true mediators (i.e., having non-zero alpha and beta coefficients).

Specifically, the first 10 coefficients of alpha are set to take some non-zero value a, the 6th to 15th coefficients of beta are set to take some non-zero value b, and all other alpha and beta coefficients are set to be 0.
}
\examples{
datagen(n = 100, K = 50, a = 0.5, b = 0.3)
}
