% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediationFDR.R
\name{MediationFDR}
\alias{MediationFDR}
\title{Mediator selection function}
\usage{
MediationFDR(
  X,
  Y,
  M,
  C1 = NULL,
  C2 = NULL,
  q1 = 0.05,
  q2 = 0.05,
  f_func = "Max",
  correction_method = "holm",
  weighted = FALSE,
  binaryOutcome = FALSE
)
}
\arguments{
\item{X}{exposure of interest}

\item{Y}{outcome of interest}

\item{M}{potential mediators}

\item{C1}{confounding variables in X-M path}

\item{C2}{confounding variables in M-Y path}

\item{q1}{FWER level in step 1}

\item{q2}{FDR controlled level in step 2}

\item{f_func}{function form of W in step 2}

\item{correction_method}{multiple comparison correction method in step 1}

\item{weighted}{whether we weight the Z statistics in step 2}

\item{binaryOutcome}{whether the outcome is binary}
}
\value{
selected mediators
}
\description{
This function selects the mediators while controlling for FDR.
}
\examples{
data = datagen(n = 200, K = 50, a = 0.8, b = 0.5)
X = data$X; Y = data$Y; M = data$M
MediationFDR(X,Y,M)

}
